Relocate record_activation_metrics()
