+++ MaxText/metric_logger.py
@@ -36,6 +36,26 @@ from MaxText.globals import EPS
 from collections import defaultdict
 
 
+def record_activation_metrics(output_metrics, intermediate_outputs, config):
+  """Adds the activation metrics to the metrics dict"""
+
+  if config.scan_layers:
+    metrics_dict = intermediate_outputs["intermediates"]["decoder"]["decoder"]
+
+    for layer_num in range(config.num_decoder_layers):
+      output_metrics["scalar"][f"activ_fraction_zero/layer_{layer_num:03d}"] = metrics_dict["activation_fraction_zero"][0][
+          layer_num
+      ]
+      output_metrics["scalar"][f"activ_mean/layer_{layer_num:03d}"] = metrics_dict["activation_mean"][0][layer_num]
+      output_metrics["scalar"][f"activ_stdev/layer_{layer_num:03d}"] = metrics_dict["activation_stdev"][0][layer_num]
+  else:
+    for layer_num in range(config.num_decoder_layers):
+      layer = intermediate_outputs["intermediates"]["decoder"][f"layers_{layer_num}"]
+      output_metrics["scalar"][f"activ_fraction_zero/layer_{layer_num:03d}"] = layer["activation_fraction_zero"][0]
+      output_metrics["scalar"][f"activ_mean/layer_{layer_num:03d}"] = layer["activation_mean"][0]
+      output_metrics["scalar"][f"activ_stdev/layer_{layer_num:03d}"] = layer["activation_stdev"][0]
+
+
 def _prepare_metrics_for_json(metrics, step, run_name):
   """Converts metric dictionary into json supported types (e.g. float)"""
   metrics_dict = {val: float(metrics["scalar"][val]) for val in metrics["scalar"]}
